requirements.txt
# requirements.txt
# Необходимое окружение python 3.11 - (python3.11 -m venv venv)
aiogram==3.8.0
aiohttp>=3.8.0
asyncio
base64
beautifulsoup4
g4f[All]
langdetect
pollinations
pillow
python-dotenv
requests
speechmatics-python